+++ packages/es-markdown-core/src/model/image.ts
@@ -1,9 +1,10 @@
 import type { MarkupBase } from "./base.js";
 import type { Markup } from "./markup.js";
+import type { Destination } from "../values.js";
 
 export interface Image extends MarkupBase<"image"> {
-    /** Required, for the reason `Link.destination` is. */
-    readonly source: string;
+    /** Required, for the reason `Link.dest` is. */
+    readonly dest: Destination;
     /** Optional. */
     readonly title: string | null;
     readonly content: readonly Markup[];